Execute usecases with context-params
