+++ shared/arkitekt-cr-usecases/build.gradle.kts
@@ -19,6 +19,10 @@ kotlin {
     iosArm64()
     iosSimulatorArm64()
 
+    compilerOptions {
+        freeCompilerArgs.add("-Xcontext-parameters")
+    }
+
     sourceSets {
         commonMain {
             dependencies {
