+++ shared/arkitekt-cr-usecases/src/commonMain/kotlin/app/futured/arkitekt/crusecases/scope/CoroutineScopeOwner.kt
@@ -1,8 +1,19 @@
 package app.futured.arkitekt.crusecases.scope
 
+import app.futured.arkitekt.crusecases.FlowUseCase
+import app.futured.arkitekt.crusecases.UseCase
+import app.futured.arkitekt.crusecases.error.UseCaseErrorHandler
+import app.futured.arkitekt.crusecases.error.UseCaseErrorHandler.globalOnErrorLogger
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlin.coroutines.cancellation.CancellationException
 
+/**
+ * This interface gives your class ability to execute [UseCase] and [FlowUseCase] Coroutine use cases.
+ * You may find handy to implement this interface in custom Presenters, ViewHolders etc.
+ * It is your responsibility to cancel [coroutineScope] when all running tasks should be stopped.
+ */
 interface CoroutineScopeOwner {
     /**
      * [CoroutineScope] scope used to execute coroutine based use cases.
@@ -15,4 +26,37 @@ interface CoroutineScopeOwner {
      * Provides Dispatcher for background tasks. This may be overridden for testing purposes.
      */
     fun getWorkerDispatcher() = Dispatchers.Default
+
+    /**
+     * Launch suspend [block] in [viewModelScope].
+     *
+     * Encapsulates this call with try catch block and when an exception is thrown
+     * then it is logged in [UseCaseErrorHandler.globalOnErrorLogger] and handled by [defaultErrorHandler].
+     *
+     * If exception is [CancellationException] then [defaultErrorHandler] is not called and
+     * [UseCaseErrorHandler.globalOnErrorLogger] is called only if the root cause of this exception is not
+     * [CancellationException] (e.g. when [Result.getOrCancel] is used).
+     */
+    @Suppress("TooGenericExceptionCaught")
+    fun launchWithHandler(block: suspend CoroutineScope.() -> Unit) {
+        viewModelScope.launch {
+            try {
+                block()
+            } catch (exception: CancellationException) {
+                val rootCause = exception.cause
+                if (rootCause != null && rootCause !is CancellationException) {
+                    UseCaseErrorHandler.globalOnErrorLogger(exception)
+                }
+            } catch (exception: Throwable) {
+                UseCaseErrorHandler.globalOnErrorLogger(exception)
+                defaultErrorHandler(exception)
+            }
+        }
+    }
+
+    /**
+     * This method is called when coroutine launched with [launchWithHandler] throws an exception and
+     * this exception isn't [CancellationException]. By default, it rethrows this exception.
+     */
+    fun defaultErrorHandler(exception: Throwable): Unit = throw exception
 }
